Rename Parameters to WsPermissionsParameters
